fix(session-turn): wrap W1 toolbar buttons in Tooltip for hover hints
AstroHan W1 retest #5: the user bubble + agent round toolbars rendered
only an `aria-label` on each action button, so pointer users had no
hover affordance explaining what the icons did. Reusing the existing
`Tooltip` component (already used by the legacy `<IconButton>` toolbar
path in `user-message-display.tsx`) restores parity while keeping the
W1 leaves context-free — the tooltip body is the same i18n-resolved
string the caller passes in via `labels.{copy,copied,reset,fork}`.

Side-effect: importing `./tooltip` from `session-turn-agent-round.tsx`
pulls in the Kobalte client-only module, which throws when the test
runner evaluates it server-side. Pure helpers
(`selectFirstAssistant`, `selectLatestAssistant`, `isInterrupted`,
`computeElapsedSec`) move into a sibling `session-turn-agent-round-helpers.ts`
so the test imports them without dragging the Kobalte chain. The
component re-exports them under the original name to keep callers
stable.
